SIRo HITL tool policy eval fix and refactor (#1479)
* init SingleAgentAccessMgr
* baselines controller refactor
* set not_done_masks to True after the first action
* removed temp --sample-random-baseline-base-vel argument
* fixed ControllerHelper.controllers instantiation
* removed unused code; added type annotations
* ignore type error
* reorganized controllers into separate files
* addressing comments
* refactored controllers' actions to be np.array
* post-rebase fix
