+++ habitat-lab/habitat/gym/gym_definitions.py
@@ -16,6 +16,7 @@ from habitat.config.default_structured_configs import (
     SimulatorSensorConfig,
     ThirdRGBSensorConfig,
 )
+from habitat.core.dataset import Dataset
 from habitat.core.environments import get_env_class
 from habitat.utils.env_utils import make_env_fn
 
@@ -47,17 +48,21 @@ def _get_env_name(cfg: "DictConfig") -> Optional[str]:
     return cfg["env_task"]
 
 
-def make_gym_from_config(config: "DictConfig", dataset=None) -> gym.Env:
+def make_gym_from_config(
+    config: "DictConfig", dataset: Optional[Dataset] = None
+) -> gym.Env:
     """
     From a habitat-lab or habitat-baseline config, create the associated gym environment.
     """
     if "habitat" in config:
         config = config.habitat
+
     env_class_name = _get_env_name(config)
     env_class = get_env_class(env_class_name)
     assert (
         env_class is not None
     ), f"No environment class with name `{env_class_name}` was found, you need to specify a valid one with env_task"
+
     return make_env_fn(env_class=env_class, config=config, dataset=dataset)
