+++ habitat-lab/habitat/utils/env_utils.py
@@ -4,10 +4,10 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from typing import TYPE_CHECKING, Type, Union
+from typing import TYPE_CHECKING, Optional, Type, Union
 
+from habitat.core.dataset import Dataset
 from habitat.core.env import Env, RLEnv
-from habitat.datasets import make_dataset
 
 if TYPE_CHECKING:
     from omegaconf import DictConfig
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 def make_env_fn(
     config: "DictConfig",
     env_class: Union[Type[Env], Type[RLEnv]],
-    dataset=None,
+    dataset: Optional[Dataset] = None,
 ) -> Union[Env, RLEnv]:
     r"""Creates an env of type env_class with specified config and rank.
     This is to be passed in as an argument when creating VectorEnv.
@@ -32,8 +32,8 @@ def make_env_fn(
     """
     if "habitat" in config:
         config = config.habitat
-    if dataset is None:
-        dataset = make_dataset(config.dataset.type, config=config.dataset)
+
     env = env_class(config=config, dataset=dataset)
     env.seed(config.seed)
+
     return env
